+++ core/src/main/scala/org/apache/spark/storage/BlockManagerMessages.scala
@@ -39,7 +39,7 @@ private[spark] object BlockManagerMessages {
   case object DecommissionBlockManager extends ToBlockManagerMasterStorageEndpoint
 
   // Remove all blocks belonging to a specific RDD.
-  case class RemoveRdd(rddId: Int) extends ToBlockManagerMasterStorageEndpoint
+  case class RemoveRdd(rddId: Long) extends ToBlockManagerMasterStorageEndpoint
 
   // Remove all blocks belonging to a specific shuffle.
   case class RemoveShuffle(shuffleId: Int) extends ToBlockManagerMasterStorageEndpoint
@@ -120,10 +120,10 @@ private[spark] object BlockManagerMessages {
 
   // Seal an RDD's per-block content checksums: per block keep the plurality checksum, evict
   // divergent copies, and reject future divergent registrations.
-  case class SealRddChecksums(rddId: Int) extends ToBlockManagerMaster
+  case class SealRddChecksums(rddId: Long) extends ToBlockManagerMaster
 
   // Invariant-check a sealed RDD; see BlockManagerMasterEndpoint.verifyRddChecksumSeal.
-  case class VerifyRddChecksumSeal(rddId: Int) extends ToBlockManagerMaster
+  case class VerifyRddChecksumSeal(rddId: Long) extends ToBlockManagerMaster
 
   case class GetLocations(blockId: BlockId) extends ToBlockManagerMaster
